Separate SFT and SFTMap from space
