+++ src/policy/largeobjectspace.rs
@@ -3,9 +3,10 @@ use atomic::Ordering;
 use crate::plan::ObjectQueue;
 use crate::plan::PlanConstraints;
 use crate::plan::VectorObjectQueue;
+use crate::policy::sft::GCWorkerMutRef;
+use crate::policy::sft::SFT;
 use crate::policy::space::SpaceOptions;
-use crate::policy::space::*;
-use crate::policy::space::{CommonSpace, Space, SFT};
+use crate::policy::space::{CommonSpace, Space};
 use crate::util::constants::BYTES_IN_PAGE;
 use crate::util::heap::layout::heap_layout::{Mmapper, VMMap};
 use crate::util::heap::HeapMeta;
