+++ src/policy/mod.rs
@@ -17,6 +17,8 @@ pub mod space;
 pub mod copy_context;
 /// Policy specific GC work
 pub mod gc_work;
+pub mod sft;
+pub mod sft_map;
 
 pub mod copyspace;
 pub mod immix;
